Generate cpu map strings for resctrl
User can input a cpu set from API as follow:
["2-8,^3-4,^7,9", "56-87,^86,^57-58"]

It can translate it to:
bffffc,1f000000,366

"2-8" means set the bit 2-8 as 1,
"^57-58" means clean the bit 57-58 to 0.

Also add some sanity check.
1. the bit index can not larger than the cpu numbers.
Such as, on a host with 88 cpus, this setting ["58-89"] exceed the
cpu numbers, this is not allowed.

2. Only allow digital(0-9), hyphen, comma, Caret characters.
["56-87, ^86,^57-58"] with a whitespace in it is also not allowed.

please ref the testcase how to use this function.