Fix: fix logic when write cpu map to resctrl
Resctrl will create cpus file with default value, so we don't need to do
anything if the cpus is equal to "" .

root@s2600wt:/sys/fs/resctrl# mkdir foo
root@s2600wt:/sys/fs/resctrl# cat foo/cpus
000000,00000000,00000000

But even this, we requires that up layer to provide cpus explicitly as
for this case, if admin want to remove cpus by performing a update, he
may pass a empty cores_ids (can't be "0", that means on cpus 0, the cpu bit
map is 0,0,1). So in this case, up layer need to pass "0" to resctrl.

Co-Authored-By: Feng, Shaohe <shaohe.feng@intel.com>
Signed-off-by: Qiao, Liyong <liyong.qiao@intel.com>