Fix: Don't create Bitmap from null
NewBitmap doesn't allow us pass null, it will cause out of range error.

When try to validate workload, if a workload has CoreIDs, RDAgent will
loop for all workload to verify if the CoreIDs are assigned to other
workload.

So there's a case that if there are workload using TaskIDs to do cache
enforce, the CoreIDs is null, this will cause NewBitmap raise runtime error:
index out of range

To avoid this, check CoreIDs length before pass it to NewBitmap.

Signed-off-by: Qiao, Liyong <liyong.qiao@intel.com>