Commit transaction.
Commit should be a transaction.
So we use taskFlow to guarantee the consistency.
Also we need a coarse granularity lock for IPC. We already has it.
Also we need a file lock for consistency among different processes.
In plan.

After some test on taskFlow, we can remove the current Commit logic
code and use taskFlow.
The taskFlow need a snapshot of all ResAssociation for the
transaction.