Workload: Support cache id awareness when do enforcement.
On a multiple cache resource host (e.g. 2 Sockets), different cpus will
be served by different cache, so for these case, if Admin specifies a
workload's cpus list, and all cpus are located on one cache, we don't
need to allocate cache on the other caches.

For a workload has:
  "core_ids": [
     "85-87
     ]

And cache topology is:

0: "share_cpu_list": "0-21,44-65"
1: "share_cpu_list": "22-43,66-87"

Only allocate cache on cache_id = 1

Signed-off-by: Qiao, Liyong <liyong.qiao@intel.com>