mini refine DB.
Though the current DB design is flat.
This patch does not want to do major operation on DB.
Just want to fix some bugs.

The DB layer is not very clear.
It is really a twist to use it.

Also need to uncoupled db and app server.

NOTE, this patch does not test Mgo.

There is also another problem, we can create multi-infra, is this
allowed?
$ curl -H "Content-Type: application/json" --request GET http://127.0.$
:8081/v1/workloads/
[
  {
   "id": "43",
   "core_ids": null,
   "task_ids": [
    "41187"
   ],
   "policy": "silver",
   "Status": "Successful",
   "group": [
    "infra"
   ],
   "CosName": "infra"
  },
  {
   "id": "44",
   "core_ids": null,
   "task_ids": [
    "41347"
   ],
   "policy": "silver",
   "Status": "Successful",
   "group": [
    "infra"
   ],
   "CosName": "infra"
  },
  {
   "id": "45",
   "core_ids": null,
   "task_ids": [
    "41469"
   ],
   "policy": "gold",
   "Status": "Successful",
   "group": [
    "infra"
   ],
   "CosName": "infra"
  }
 ]