Shrink cache in besteffort pool
We allocate as more cache ways as we can between MinCache and MaxCache.
If there's not enough cache way left in Best effort pool, and we still
want to allocate cache in Best effort pool, we need to shrink the
allocated cache ways for workloads.

This patch implement a simple shrink method:

If the host can not satisfy MinCache for a best effort pool request, try to
shrink already allocated cache ways in best effort pool and allocate the
MinCache for the new request. Allocate cache ways from low bit, so it won't
waste cache ways.

Signed-off-by: Qiao, Liyong <liyong.qiao@intel.com>